pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "book-codex-agent"
version = "0.1.0"
description = "Custom Codex-style Book Agent: LangChain + ChatGPT + Git/GitHub + Typer CLI"
readme = "README.md"
authors = [{name="Your Name", email="you@example.com"}]
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12.3",
    "rich>=13.7.1",
    "python-dotenv>=1.0.1",
    "gitpython>=3.1.43",
    "PyGithub>=2.3.0",
    "langchain>=0.2.11",
    "openai>=1.40.2",
    "pydantic>=2.8.2",
    "pyyaml>=6.0.2"
]

[project.scripts]
book-agent = "book_agent.cli:app"
